DAL -> MVC straight connection
