--- BusinessLogic/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace BusinessLogic
-{
-    public class Class1
-    {
-    }
-}
