+++ BusinessLogic/Interfaces/IPage.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BusinessLogic.Interfaces
+{
+    class IPage
+    {
+        
+    }
+}
