+++ BusinessLogic/Interfaces/IPageContainer.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BusinessLogic.Interfaces
+{
+    public interface IPageContainer
+    {
+        void GetAllText();
+    }
+}
